Fix: correct format specifier mismatches caught by the new attribute
- DumpCurResGroup passed a size_t element count to %d (use %zu)
- LawnGetCurrentLevelName passed a C string to %d, printing pointer
  digits as a garbage level number in loading logs for adventure
  replays (use F%s for the intended F<stage> text)
- the session id log passed an intptr_t to %u (use %lld with a cast)
- DataArrayFree/DataArrayGetID's asserts passed an object pointer to
  0x%x, truncating it on 64-bit (use %p)

+++ src/LawnApp.cpp
@@ -1233,7 +1233,7 @@ void LawnApp::Init()
 	// these debug checks break the whole exe in release mode
 //#ifdef PVZ_DEBUG
 	PvzpAssertInitForApp();
-	PvzpLogLn("session id: %u", mSessionID);
+	PvzpLogLn("session id: %lld", static_cast<long long>(mSessionID));
 //#endif
 
 	if (!mResourceManager->ParseResourcesFile("properties/resources.xml"))
